+++ apps/web/src/app/(web)/[slug]/page.tsx
@@ -1,2 +1,0 @@
-export * from '@/page/GenericPage'
-export {default} from '@/page/GenericPage'