+++ apps/web/src/app/(web)/blog/[slug]/page.tsx
@@ -1,2 +1,0 @@
-export * from '@/page/BlogPostPage'
-export {default} from '@/page/BlogPostPage'