+++ apps/web/src/app/(web)/docs/[framework]/[[...slug]]/page.tsx
@@ -1,2 +1,0 @@
-export * from '@/page/DocPage'
-export {default} from '@/page/DocPage'