+++ apps/web/src/app/page.tsx
@@ -1,2 +1,0 @@
-export * from '@/page/HomePage'
-export {default} from '@/page/HomePage'