+++ apps/web/src/page/demo/DemoDashboard.dynamic.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-import dynamic from 'next/dynamic'
-
-export const DemoPage = dynamic(() => import('./DemoDashboard'), {ssr: false})