+++ src/adapter/next/cms.tsx
@@ -1,19 +1,15 @@
 import {Headers} from '@alinea/iso'
-import {createPreviewParser} from 'alinea/backend/resolver/ParsePreview'
-import {generatedSource} from 'alinea/backend/store/GeneratedSource'
 import {COOKIE_NAME} from 'alinea/cloud/CloudRemote'
-import {CMS} from 'alinea/core/CMS'
 import {Client} from 'alinea/core/Client'
+import {CMS} from 'alinea/core/CMS'
 import type {Config} from 'alinea/core/Config'
 import type {UploadResponse} from 'alinea/core/Connection'
+import type {Mutation} from 'alinea/core/db/Mutation'
 import type {GraphQuery} from 'alinea/core/Graph'
 import {outcome} from 'alinea/core/Outcome'
 import type {PreviewRequest} from 'alinea/core/Preview'
 import type {User} from 'alinea/core/User'
-import {LocalDB} from 'alinea/core/db/LocalDB'
-import type {Mutation} from 'alinea/core/db/Mutation'
 import {getPreviewPayloadFromCookies} from 'alinea/preview/PreviewCookies'
-import PLazy from 'p-lazy'
 import {requestContext} from './context.js'
 
 export interface PreviewProps {
@@ -25,35 +21,13 @@ export interface PreviewProps {
 export class NextCMS<
   Definition extends Config = Config
 > extends CMS<Definition> {
-  lastSync = 0
-
-  db = PLazy.from(async () => {
-    if (process.env.NEXT_RUNTIME === 'edge')
-      throw new Error('Local db not available in edge')
-    const source = await generatedSource
-    const db = new LocalDB(this.config, source)
-    await db.sync()
-    return db
-  })
-
-  init = PLazy.from(async () => {
-    const db = await this.db
-    const previews = createPreviewParser(db)
-    return {db, previews}
-  })
-
   constructor(config: Definition) {
     super(config)
   }
 
-  async sync() {
-    const {db} = await this.init
-    return db.sync()
-  }
-
   async resolve<Query extends GraphQuery>(query: Query): Promise<any> {
     let status = query.status
-    const {isDev, handlerUrl, apiKey} = await requestContext(this.config)
+    const {handlerUrl, apiKey} = await requestContext(this.config)
     const client = new Client({
       config: this.config,
       url: handlerUrl.href,
@@ -72,25 +46,7 @@ export class NextCMS<
       const payload = getPreviewPayloadFromCookies(cookie.getAll())
       if (payload) preview = {payload}
     }
-    if (process.env.NEXT_RUNTIME === 'edge' || isDev)
-      return client.resolve({preview, ...query, status})
-    const {PHASE_PRODUCTION_BUILD} = await import('next/constants')
-    const isBuild = process.env.NEXT_PHASE === PHASE_PRODUCTION_BUILD
-    const {db, previews} = await this.init
-    const sync = () => db.syncWith(client).catch(console.error)
-    if (!isBuild) {
-      if (preview) {
-        preview = await previews.parse(preview, sync)
-      } else {
-        const syncInterval = query.syncInterval ?? 60
-        const now = Date.now()
-        if (now - this.lastSync >= syncInterval * 1000) {
-          this.lastSync = now
-          await sync()
-        }
-      }
-    }
-    return db.resolve({preview, ...query, status})
+    return client.resolve({preview, ...query, status})
   }
 
   async #authenticatedClient() {