+++ src/core/source/IndexedDBSource.ts
@@ -1,14 +1,29 @@
 import type {ChangesBatch} from './Change.js'
+import {ShaMismatchError} from './ShaMismatchError.js'
 import type {Source} from './Source.js'
 import {ReadonlyTree} from './Tree.js'
 
 export class IndexedDBSource implements Source {
-  #db: Promise<IDBDatabase>
+  #factory: IDBFactory
+  #name: string
+  #connection: Promise<IDBDatabase> | undefined
 
   constructor(indexedDB: IDBFactory, name: string) {
-    this.#db = new Promise<IDBDatabase>((resolve, reject) => {
-      const request = indexedDB.open(name)
-      request.onsuccess = () => resolve(request.result)
+    this.#factory = indexedDB
+    this.#name = name
+  }
+
+  #createConnection(): Promise<IDBDatabase> {
+    return new Promise<IDBDatabase>((resolve, reject) => {
+      const request = this.#factory.open(this.#name)
+      request.onsuccess = () => {
+        const db = request.result
+        db.onclose = () => {
+          console.info('IndexedDB connection closed')
+          this.#connection = undefined
+        }
+        resolve(db)
+      }
       request.onerror = () => reject(request.error)
       request.onupgradeneeded = () => {
         const db = request.result
@@ -18,18 +33,52 @@ export class IndexedDBSource implements Source {
     })
   }
 
-  async getTree() {
-    const db = await this.#db
-    return new Promise<ReadonlyTree>((resolve, reject) => {
-      const transaction = db.transaction(['tree'], 'readonly')
-      const store = transaction.objectStore('tree')
-      const request = store.get('tree')
-      request.onsuccess = event => {
-        const entry = (event.target as IDBRequest).result
-        resolve(entry ? new ReadonlyTree(entry) : ReadonlyTree.EMPTY)
+  #connect(): Promise<IDBDatabase> {
+    this.#connection ??= this.#createConnection()
+    return this.#connection
+  }
+
+  #retryIfClosing<T extends Function>(handle: T) {
+    return (error: Error) => {
+      if (error instanceof Error && error.message.includes('closing')) {
+        // If the database is closing, we retry the operation handle
+        this.#connection = undefined
+        return handle()
       }
-      request.onerror = event => reject((event.target as IDBRequest).error)
-    })
+      throw error
+    }
+  }
+
+  getTree(): Promise<ReadonlyTree> {
+    const handle = async () => {
+      const db = await this.#connect()
+      const transaction = db.transaction(['tree', 'blobs'], 'readonly')
+      const treeStore = transaction.objectStore('tree')
+      const blobsStore = transaction.objectStore('blobs')
+      const [tree, blobKeys] = await Promise.all([
+        new Promise<ReadonlyTree>((resolve, reject) => {
+          const request = treeStore.get('tree')
+          request.onsuccess = event => {
+            const entry = (event.target as IDBRequest).result
+            resolve(entry ? new ReadonlyTree(entry) : ReadonlyTree.EMPTY)
+          }
+          request.onerror = event => reject((event.target as IDBRequest).error)
+        }),
+        new Promise<Array<string>>((resolve, reject) => {
+          const request = blobsStore.getAllKeys()
+          request.onsuccess = () => resolve(request.result as Array<string>)
+          request.onerror = event => reject((event.target as IDBRequest).error)
+        })
+      ])
+      for (const sha of tree.shas) {
+        if (!blobKeys.includes(sha)) {
+          console.warn(`Blob ${sha} in tree, but not found`)
+          return ReadonlyTree.EMPTY
+        }
+      }
+      return tree
+    }
+    return handle().catch(this.#retryIfClosing(handle))
   }
 
   async getTreeIfDifferent(sha: string): Promise<ReadonlyTree | undefined> {
@@ -40,7 +89,7 @@ export class IndexedDBSource implements Source {
   async *getBlobs(
     shas: Array<string>
   ): AsyncGenerator<[sha: string, blob: Uint8Array]> {
-    const db = await this.#db
+    const db = await this.#connect()
     const transaction = db.transaction(['blobs'], 'readonly')
     const store = transaction.objectStore('blobs')
     for (const sha of shas) {
@@ -57,8 +106,14 @@ export class IndexedDBSource implements Source {
   }
 
   async applyChanges(batch: ChangesBatch) {
-    const db = await this.#db
+    const db = await this.#connect()
     const current = await this.getTree()
+    if (batch.fromSha !== current.sha)
+      throw new ShaMismatchError(
+        current.sha,
+        batch.fromSha,
+        'Cannot apply changes locally due to SHA mismatch'
+      )
     const updatedTree = current.clone()
     updatedTree.applyChanges(batch)
     const compiled = await updatedTree.compile()
@@ -71,10 +126,15 @@ export class IndexedDBSource implements Source {
         case 'add':
           blobs.put(change.contents, change.sha)
           break
-        case 'delete':
-          blobs.delete(change.sha)
-          break
       }
+    const blobKeys = await new Promise<Array<string>>((resolve, reject) => {
+      const request = blobs.getAllKeys()
+      request.onsuccess = () => resolve(request.result as Array<string>)
+      request.onerror = event => reject((event.target as IDBRequest).error)
+    })
+    for (const sha of blobKeys) {
+      if (!compiled.hasSha(sha)) blobs.delete(sha)
+    }
     return new Promise<void>((resolve, reject) => {
       transaction.oncomplete = () => resolve()
       transaction.onerror = event => reject((event.target as IDBRequest).error)